refactor: split ide.js, remove variable scope and syntax checker functionality
