+++ app/scripts/controllers/ide.js
@@ -20,11 +20,8 @@ app.controller('IdeCtrl', [
   'UserSrv',
   'WebsocketSrv',
   'ChatSrv',
-  'CodingAssistantCodeMatchSrv',
   'CodeboardSrv',
-  'AceEditorSrv',
   'AISrv',
-  'UITexts',
   function (
     $scope,
     $rootScope,
@@ -43,11 +40,8 @@ app.controller('IdeCtrl', [
     UserSrv,
     WebsocketSrv,
     ChatSrv,
-    CodingAssistantCodeMatchSrv,
     CodeboardSrv,
-    AceEditorSrv,
-    AISrv,
-    UITexts
+    AISrv
   ) {
     // First we handle all data that was injected as part of the app.js resolve.
     // set the project data in the ProjectFactory (files, description, etc.)
@@ -1150,7 +1144,6 @@ app.controller('IdeCtrl', [
       tool: false,
       submit: false,
       beautify: false,
-      varScope: false,
     };
 
     // state variables to indicate which actions in the IDE are hidden
@@ -1357,7 +1350,7 @@ app.controller('IdeCtrl', [
     /**
      * prepare panes depending on hidden actions (Janick Michot)
      */
-    var treeViewOpen = $scope.currentRoleIsOwner() || $scope.currentRoleIsHelp() || $scope.currentRoleIsSubmission();
+    const treeViewOpen = $scope.currentRoleIsOwner() || $scope.currentRoleIsHelp() || $scope.currentRoleIsSubmission();
     // kPane for treeView
     $scope.kPanes = $scope.isActionHidden('tree-view')
       ? '['
@@ -1512,11 +1505,6 @@ app.controller('IdeCtrl', [
             syncEditor();
             formatCode();
           });
-          break;
-        case 'show_var_scope':
-          CodingAssistantCodeMatchSrv.toggleMarkers($scope.ace.editor, false);
-          $scope.variableMap = CodeboardSrv.getVariableMap();
-          $scope.toggleVarScope();
           break;
       }
     };
@@ -2040,877 +2028,10 @@ app.controller('IdeCtrl', [
     };
 
     /** Below list all one-time invocations for functions which should run whenever the controller is loaded from scratch.*/
-
     // invoke function to init the project
     initProject();
     (async () => {
       await $scope.checkRemainingAIRequests();
     })();
-
-    /////////////////////////////////////////////// syntax-checker and variable scope functionality (part of coding-assistant) ///////////////////////////////////////////////
-    var disabledActions = CodeboardSrv.getDisabledActions();
-    var enabledActions = CodeboardSrv.getEnabledActions();
-    var errorLine;
-    var currentLine;
-
-    // fetch db data from CodingAssistantCodeMatchSrv
-    function fetchData() {
-      return CodingAssistantCodeMatchSrv.getJsonData()
-        .then((db) => {
-          return { db };
-        })
-        .catch((error) => {
-          console.error('An error occurred while fetching data:', error);
-        });
-    }
-
-    fetchData().then(({ db }) => {
-      // call updateExplanations once to ensure that initial code is loaded
-      updateExplanations(db);
-
-      // call updateExplanations() with a slight delay to ensure the initial code is loaded (click file in tree-view case)
-      $scope.$on(IdeMsgService.msgJavaFileOpened().msg, function () {
-        $timeout(() => {
-          updateExplanations(db);
-        });
-      });
-
-      // call updateExplanations() with a slight delay to ensure the initial code is loaded (click tab above editor case)
-      $scope.$on(IdeMsgService.msgJavaTabOpened().msg, function () {
-        $timeout(() => {
-          updateExplanations(db);
-        });
-      });
-
-      // call updateExplanations() with a slight delay to ensure the code is loaded (reset code case)
-      $scope.$on(IdeMsgService.msgResetCode().msg, function () {
-        $timeout(() => {
-          updateExplanations(db);
-        });
-      });
-
-      AceEditorSrv.aceChangeListener($scope.ace.editor, function () {
-        var lSelectedNode = CodeboardSrv.getFile() || '.java';
-        if (lSelectedNode.match(/.java/)) {
-          // automatically call $apply if necessarry to prevent '$apply already in progress' error
-          $timeout(() => {
-            updateExplanations(db);
-          });
-          // broadcast that varScope window gets closed when there was a change in the code...
-          $rootScope.$broadcast(IdeMsgService.msgCodeChanged().msg);
-        }
-
-        // add markers dynamically
-        // CodingAssistantCodeMatchSrv.addDynamicMarkers(aceEditor););
-      });
-    });
-
-    function updateExplanations(db) {
-      var annotations = [];
-      var inputCode = AceEditorSrv.getInputCode($scope.ace.editor);
-      var result = CodingAssistantCodeMatchSrv.getMatchedExplanations(db, inputCode, $scope.ace.editor);
-      // convert variableMap into an object
-      CodeboardSrv.setVariableMap(Object.fromEntries(result.variableMap));
-
-      result.explanations.forEach((explanation) => {
-        if (explanation.isError) {
-          // lineLevel of error chatbox
-          errorLine = explanation.lineLevel;
-          // current lineLevel of cursor
-          currentLine = $scope.ace.editor.getSelectionRange().start.row + 1;
-          if (currentLine !== errorLine) {
-            // store a new annotation with the error lineLevel in the annotations array
-            annotations.push({
-              row: explanation.lineLevel - 1,
-              column: 0,
-              text: explanation.answer,
-              type: 'error',
-            });
-          }
-        }
-        if (!disabledActions.includes('syntax-checker') || enabledActions.includes('syntax-checker')) {
-          // display all the annotations in the aceEditor
-          $scope.ace.editor.getSession().setAnnotations(annotations);
-        }
-      });
-    }
-  },
-]);
-
-app.controller('TreeCtrl', [
-  '$scope',
-  '$rootScope',
-  '$log',
-  'ProjectFactory',
-  'IdeMsgService',
-  'CodeboardSrv',
-  function ($scope, $rootScope, $log, ProjectFactory, IdeMsgService, CodeboardSrv) {
-    $scope.projectNodes = ProjectFactory.getProject().files;
-
-    /**
-     * Adds a new file to the project model (which feeds the tree view)
-     */
-    $scope.addFile = function () {
-      var req = IdeMsgService.msgNewNodeRequest('file');
-      $rootScope.$broadcast(req.msg, req.data);
-    };
-
-    /**
-     * Adds a new folder to the project model (which feeds the tree view)
-     */
-    $scope.addFolder = function () {
-      var req = IdeMsgService.msgNewNodeRequest('folder');
-      $rootScope.$broadcast(req.msg, req.data);
-    };
-
-    /**
-     * Broadcasts a msg that a node was selected (only if selected node is not a folder).
-     */
-    $scope.nodeClick = function () {
-      // broadcast an event when a file is opened
-      var lSelectedNode = ProjectFactory.getNode($scope.mytree.currentNode.uniqueId);
-      CodeboardSrv.setFile(lSelectedNode.filename);
-      // only broadcast an event when a java file is opened (click file in tree-view case)
-      if (lSelectedNode.filename.match(/.java/)) {
-        $rootScope.$broadcast(IdeMsgService.msgJavaFileOpened().msg);
-      }
-
-      // ignore the click if the selected node is a folder
-      if (lSelectedNode !== null && !lSelectedNode.isFolder) {
-        var req = IdeMsgService.msgDisplayFileRequest(lSelectedNode.uniqueId);
-        $rootScope.$broadcast(req.msg, req.data);
-      }
-    };
-
-    /**
-     * Listen for an event to rename a node. The currently selected node will be renamed.
-     */
-    $scope.$on(IdeMsgService.msgRenameNodeRequest().msg, function (aEvent) {
-      // if the node to rename is the root node, we prevent it
-      if ($scope.mytree.currentNode.uniqueId === 0) {
-        alert("The root folder can't be renamed.");
-      } else {
-        // get the file name and file type
-        var lNodeId = $scope.mytree.currentNode.uniqueId;
-        var lNodeName = ProjectFactory.getNode(lNodeId).filename;
-        var lNodeType = ProjectFactory.getNode(lNodeId).isFolder ? 'folder' : 'file';
-
-        var req = IdeMsgService.msgDisplayRenameNodeModalRequest(lNodeId, lNodeName, lNodeType);
-        $rootScope.$broadcast(req.msg, req.data);
-      }
-    });
-
-    /**
-     * Listens for the event when the user has provided a new name for the node that should be renamed.
-     */
-    $scope.$on(IdeMsgService.msgRenameNodeNameAvailable().msg, function (aEvent, aMsgData) {
-      ProjectFactory.renameNode(aMsgData.nodeId, aMsgData.nodeName);
-
-      //$scope.projectNodes = ProjectFactory.getProject().files;
-
-      // need to broadcast a ReloadTreeFromProjectFactory msg
-      $rootScope.$broadcast(IdeMsgService.msgReloadTreeFromProjectFactory().msg);
-    });
-
-    /**
-     * Listens for an event to remove a node. The currently selected node will be removed.
-     */
-    $scope.$on(IdeMsgService.msgRemoveNodeRequest().msg, function (aEvent) {
-      // if the node to delete is the root node, we prevent it
-      if ($scope.mytree.currentNode.uniqueId === 0) {
-        alert("The root folder can't be deleted.");
-      } else {
-        // get the file name
-        var filename = ProjectFactory.getNode($scope.mytree.currentNode.uniqueId).filename;
-        var type = ProjectFactory.getNode($scope.mytree.currentNode.uniqueId).isFolder ? 'folder' : 'file';
-
-        var confirmMsg = 'Do you really want to delete ' + type + " '" + filename + "'?";
-        var confirmMsg = ProjectFactory.getNode($scope.mytree.currentNode.uniqueId).isFolder
-          ? confirmMsg + "\n\nNote: when deleting a folder, make sure it's empty."
-          : confirmMsg;
-
-        // ask the user to confirm deletion.
-        var _userConfirmed = confirm(confirmMsg);
-        if (_userConfirmed) {
-          var lSelectedNodeUId = $scope.mytree.currentNode.uniqueId;
-          ProjectFactory.removeNode(lSelectedNodeUId, false);
-
-          // Note: we need to select a new node; that's gonna be the root node
-          // set the root node to be selected
-          ProjectFactory.getNode(0).selected = 'selected';
-          // set the root node as the current node
-          $scope.mytree.currentNode = ProjectFactory.getNode(0);
-
-          // broadcast a message about which node was removed; e.g. tabs belonging to this node need to be closed
-          var req = IdeMsgService.msgRemoveNodeConfirmation(lSelectedNodeUId);
-          $rootScope.$broadcast(req.msg, req.data);
-        }
-      }
-    });
-
-    /**
-     * Listens for the event when the user has provided a name for the new node that should be added.
-     */
-    $scope.$on(IdeMsgService.msgNewNodeNameAvailable().msg, function (aEvent, aMsgData) {
-      var lSelectedNodeUId = $scope.mytree.currentNode.uniqueId;
-
-      switch (aMsgData.nodeType) {
-        case 'file':
-          ProjectFactory.addFile(lSelectedNodeUId, aMsgData.nodeName);
-          break;
-        case 'folder':
-          ProjectFactory.addFolder(lSelectedNodeUId, aMsgData.nodeName);
-          break;
-      }
-    });
-
-    /**
-     * Listens for the event that the tree should update.
-     * Such an update is needed if the ProjectFactory changes the 'files' array (e.g.
-     * when loading a user's version of a project) after the 'files' array was already
-     * assigned to $scope.projectNodes
-     */
-    $scope.$on(IdeMsgService.msgReloadTreeFromProjectFactory().msg, function (aEvent) {
-      // update the scope to reference the new version of the project
-      $scope.projectNodes = ProjectFactory.getProject().files;
-    });
-
-    /**
-     * Listens for the event that node should be hidden.
-     * If received the currently selected node will be hidden or unhidden, depending on its current status.
-     */
-    $scope.$on(IdeMsgService.msgHideNodeRequest().msg, function () {
-      $log.debug('Hide_node request received');
-
-      var lSelectedNodeUId = $scope.mytree.currentNode.uniqueId;
-
-      if (lSelectedNodeUId > 0) {
-        $log.debug('Node is hidden: ' + ProjectFactory.getNode(lSelectedNodeUId).isHidden);
-        ProjectFactory.setNodeHidden(ProjectFactory.getNode(lSelectedNodeUId));
-        $log.debug('Node is hidden: ' + ProjectFactory.getNode(lSelectedNodeUId).isHidden);
-      } else {
-        console.log("The root folder can't be hidden.");
-      }
-    });
-
-    /**
-     * Listens for the event that node should be uneditable.
-     * If received the currently selected node will be uneditable or editable, depending on its current status.
-     * @author Janick Michot
-     */
-    $scope.$on(IdeMsgService.msgMakeNodeStaticNodeRequest().msg, function () {
-      $log.debug('Uneditable_node request received');
-
-      var lSelectedNodeUId = $scope.mytree.currentNode.uniqueId;
-
-      if (lSelectedNodeUId > 0) {
-        ProjectFactory.setNodeStatic(ProjectFactory.getNode(lSelectedNodeUId));
-        $log.debug('Node is hidden: ' + ProjectFactory.getNode(lSelectedNodeUId).isHidden);
-      } else {
-        console.log("The root folder can't be uneditable.");
-      }
-    });
-  },
-]);
-
-app.controller('TabCtrl', [
-  '$scope',
-  '$rootScope',
-  '$log',
-  '$uibModal',
-  'ProjectFactory',
-  'IdeMsgService',
-  'CodeboardSrv',
-  function ($scope, $rootScope, $log, $uibModal, ProjectFactory, IdeMsgService, CodeboardSrv) {
-    $scope.tabs = [];
-
-    /**
-     * Function to set a particular tab as active and set all other tabs inactive.
-     * @param aArrayIndex the index of the tab to set active
-     */
-    var makeTabActive = function (aArrayIndex) {
-      for (var i = 0; i < $scope.tabs.length; i++) {
-        $scope.tabs[i].isActive = i === aArrayIndex;
-      }
-    };
-
-    /**
-     * Function to check if a tab already exists for a specific node.
-     * @param {number} aNodeId the uniqueId of the node
-     * @return {number} returns -1 if no tab exists yet, otherwise the array index of the tab
-     */
-    var doesTabAlreadyExist = function (aNodeId) {
-      for (var i = 0; i < $scope.tabs.length; i++) {
-        if ($scope.tabs[i].nodeIndex === aNodeId) {
-          return i;
-        }
-      }
-      return -1;
-    };
-
-    /**
-     * Function that is called when a tab is clicked on by the user.
-     * @param {number} aArrayIndex the array index (it's a property of a tab) of the tab that was clicked on
-     */
-    $scope.selectClick = function (aArrayIndex) {
-      var req = IdeMsgService.msgDisplayFileRequest($scope.tabs[aArrayIndex].nodeIndex);
-      $rootScope.$broadcast(req.msg, req.data);
-
-      CodeboardSrv.setFile($scope.tabs[aArrayIndex].name);
-      if ($scope.tabs[aArrayIndex].name.match(/.java/)) {
-        // broadcast an event when a new tab gets clicked (above the editor) and only when it is a .java file
-        $rootScope.$broadcast(IdeMsgService.msgJavaTabOpened().msg);
-      }
-    };
-
-    /**
-     * Removes a tab from the array
-     * @param aArrayIndex the array index at which position the tab is stored
-     */
-    $scope.closeClick = function (aArrayIndex) {
-      $log.debug('Close-tab clicked; array-index:' + aArrayIndex);
-
-      // all tabs are to the right of the tab to-be-deleted need to update their arrayIndex position
-      for (var i = aArrayIndex + 1; i < $scope.tabs.length; i++) {
-        $scope.tabs[i].arrayIndex = $scope.tabs[i].arrayIndex - 1;
-      }
-
-      // if the tab to be closed was active, pick a neighboring tab be active
-      if ($scope.tabs[aArrayIndex].isActive) {
-        // need to handle different cases:
-        if ($scope.tabs.length === 1) {
-          // the tab to-be-removed is the only tab; after closing there are not tabs open
-          // signal that the editor should be hidden
-          var req = IdeMsgService.msgDisplayEditorRequest(false);
-          $rootScope.$broadcast(req.msg, req.data);
-        } else if (aArrayIndex === 0) {
-          // other tabs exist and the tab to-be-removed is the left-most; so we activate this neighbor to the right
-          var req = IdeMsgService.msgDisplayFileRequest($scope.tabs[aArrayIndex + 1].nodeIndex);
-          $rootScope.$broadcast(req.msg, req.data);
-        } else {
-          // other tabs exist and the tab to-be-removed has neighboring tabs to the left; we select the left neighbor
-          var req = IdeMsgService.msgDisplayFileRequest($scope.tabs[aArrayIndex - 1].nodeIndex);
-          $rootScope.$broadcast(req.msg, req.data);
-        }
-      }
-
-      // remove the tab
-      $scope.tabs.splice(aArrayIndex, 1);
-    };
-
-    $scope.$on(IdeMsgService.msgDisplayFileRequest().msg, function (aEvent, aMsgData) {
-      // if no tab is displayed yet, send a message that the editor shall be displayed
-      if ($scope.tabs.length === 0) {
-        var req = IdeMsgService.msgDisplayEditorRequest(true);
-        $rootScope.$broadcast(req.msg, req.data);
-      }
-
-      // check if the requested file already has an opened tab
-      var k = doesTabAlreadyExist(aMsgData.nodeId);
-
-      if (k === -1) {
-        // there's no tab for the request file yet, so we create one
-
-        // get the node for which a tab should be added
-        var lRequestedNode = ProjectFactory.getNode(aMsgData.nodeId);
-
-        // push a new tab to the list of tabs
-        $scope.tabs.push({
-          name: lRequestedNode.filename,
-          title: lRequestedNode.path + '/' + lRequestedNode.filename,
-          nodeIndex: aMsgData.nodeId,
-          arrayIndex: $scope.tabs.length,
-          isActive: false,
-          isStatic: lRequestedNode.isStatic && !$scope.currentRoleIsOwner() ? true : false,
-        });
-
-        //make the new tab active (this call will also make the previous active tab inactive)
-        makeTabActive($scope.tabs.length - 1);
-      } else {
-        // there's already a tab opened for the request file
-        // so we simply activate that tab
-        makeTabActive(k);
-      }
-    });
-
-    /**
-     * Listens for the event when the user has provided a new name for the node that should be renamed.
-     * This event handler then changes the name of the open tab (if one exists).
-     */
-    $scope.$on(IdeMsgService.msgRenameNodeNameAvailable().msg, function (aEvent, aMsgData) {
-      // if a tab for the node-to-be-renamed is open, get its id
-      var tabId = doesTabAlreadyExist(aMsgData.nodeId);
-
-      if (tabId !== -1) {
-        $scope.tabs[tabId].name = aMsgData.nodeName;
-      }
-    });
-
-    /**
-     * Listen for a confirmation that a node was removed. If a tab for that node is open
-     * we need to close it.
-     * TODO: if remove is a folder, close all children
-     */
-    $scope.$on(IdeMsgService.msgRemoveNodeConfirmation().msg, function (aEvent, aMsgData) {
-      // if a tab for the node-to-be-removed is open, get its id
-      var tabId = doesTabAlreadyExist(aMsgData.uniqueId);
-
-      if (tabId !== -1) {
-        $scope.closeClick(tabId);
-      }
-    });
-
-    /** Handles the event that the Modal for "Share Project" should show be shown. */
-    $scope.$on(IdeMsgService.msgShowShareProjectModalRequest().msg, function (aEvent, aMsgData) {
-      // need a reference to the tabs which is accessible inside the modal; $scope.tabs won't work because the modal has it's own scope.
-      var tabs = $scope.tabs;
-
-      /** The controller for the modal */
-      var shareProjectModalInstanceCtrl = [
-        '$rootScope',
-        '$scope',
-        '$location',
-        '$uibModalInstance',
-        function ($rootScope, $scope, $location, $uibModalInstance) {
-          /** Function returns the full Url but with all query strings removed, i.e. after the '?' */
-          var getAbsUrlWithoutQueryString = function getAbsUrlWithoutQueryString() {
-            var result = $location.absUrl();
-            var queryStartIndex = result.indexOf('?');
-
-            if (queryStartIndex >= 0) {
-              result = result.substr(0, queryStartIndex);
-            }
-
-            return result;
-          };
-
-          // data-binding for the from that shows the Share-Url and a checkbox
-          // Note: we use getAbsUrlWithoutQueryString because the user might have open e.g. /projects/11?view=2.1
-          // Now if were to simply use $location.absUrl, then we would append another ?view=x.x onto the existing one, giving us /project/11?view=2.1?view=x.x
-          // To avoid this, we get the Url without the query string.
-          $scope.form = {
-            inputText: getAbsUrlWithoutQueryString(),
-            inputCheckbox: false,
-          };
-
-          /** Append or remove the "?view=..." query string */
-          $scope.checkboxChanged = function () {
-            // we always show the default, absolute Url
-            $scope.form.inputText = getAbsUrlWithoutQueryString();
-
-            // append the view query string if the checkbox is selected
-            if ($scope.form.inputCheckbox) {
-              // by default the view query string is empty
-              var viewQueryString = '';
-
-              // if tabs are open, we append to the veiw query string
-              if (tabs.length > 0) {
-                // the key of the query string
-                viewQueryString += '?view=';
-                // calculates the values of the query string
-                for (var i = 0; i < tabs.length; i++) {
-                  // get the nodeId and if the node is active or not
-                  viewQueryString += tabs[i].nodeIndex + '.' + (tabs[i].isActive ? '1' : '0');
-
-                  // add the separator "-", as in 2.1-3.0-4.0 (except after the last tab)
-                  viewQueryString += i < tabs.length - 1 ? '-' : '';
-                }
-              }
-
-              // append the calculated view query string
-              $scope.form.inputText += viewQueryString;
-            }
-          };
-
-          $scope.closeModal = function () {
-            $uibModalInstance.close();
-          };
-        },
-      ];
-
-      // call the function to open the modal (we ignore the modalInstance returned by this call as we don't need to access any data from the modal)
-      $uibModal.open({
-        templateUrl: 'ideShareProjectModal.html',
-        controller: shareProjectModalInstanceCtrl,
-      });
-    });
-  },
-]);
-
-/**
- * Controller for the new right bar
- * @author Janick Michot
- */
-app.controller('RightBarCtrl', [
-  '$scope',
-  '$rootScope',
-  'ProjectFactory',
-  'IdeMsgService',
-  'TabService',
-  'CodingAssistantCodeMatchSrv',
-  function ($scope, $rootScope, ProjectFactory, IdeMsgService, TabService, CodingAssistantCodeMatchSrv) {
-    $scope.navBarRightContent = '';
-    $scope.activeTab = '';
-    $scope.rightBarTabs = {};
-    $scope.showRightBarTabs = true;
-    $scope.isCollapsed = true;
-    var tablinks = document.getElementsByClassName('tab');
-
-    // In the following all tabs are defined, which are displayed in the right bar. The definition consists of a title,
-    // an icon and the ContentUrl. The ContentUrl specifies which template is to be loaded. These templates can in turn
-    // be controlled by a controller. With 'disable' you can also define whether the tab should be displayed. This value
-    // can be adjusted in a controller via broadcast.
-
-    // tab for project description
-    if (ProjectFactory.getProjectDescription() !== '') {
-      $scope.rightBarTabs.description = {
-        slug: 'description',
-        title: 'Aufgabe',
-        disabled: false,
-        icon: 'glyphicon-education',
-        contentURL: 'partials/navBarRight/navBarRightProjectDescription',
-      };
-    }
-
-    // tab for project description
-    if (!$scope.isActionHidden('info')) {
-      $scope.rightBarTabs.info = {
-        slug: 'info',
-        title: 'Info',
-        disabled: false,
-        icon: 'glyphicon-info-sign',
-        contentURL: 'partials/navBarRight/navBarRightInfo',
-      };
-    }
-
-    // tab for test result
-    if (!$scope.isActionHidden('test') && ProjectFactory.hasConfig('Testing', 'ioTests')) {
-      $scope.rightBarTabs.test = {
-        slug: 'test',
-        title: 'Test',
-        icon: 'glyphicon-list-alt',
-        contentURL: 'partials/navBarRight/navBarRightTest',
-      };
-    }
-
-    // tab for code explanation (coding-assistant)
-    if (!$scope.isActionHidden('explanation')) {
-      $scope.rightBarTabs.explanation = {
-        slug: 'explanation',
-        title: 'Erklärungen',
-        disabled: false,
-        icon: 'glyphicon-comment',
-        contentURL: 'partials/navBarRight/navBarRightExplanation',
-      };
-    }
-
-    // tab for compiler messages
-    if (!$scope.isActionHidden('compiler')) {
-      $scope.rightBarTabs.compiler = {
-        slug: 'compiler',
-        title: 'Compiler',
-        disabled: false,
-        icon: 'glyphicon-exclamation-sign',
-        contentURL: 'partials/navBarRight/navBarRightCompiler',
-      };
-    }
-
-    // tab for tips
-    if (!$scope.isActionHidden('tips')) {
-      $scope.rightBarTabs.tips = {
-        slug: 'tips',
-        title: 'Tipps',
-        disabled: false,
-        icon: 'glyphicon-gift',
-        contentURL: 'partials/navBarRight/navBarRightHints',
-      };
-    }
-
-    // tab for questions and answers
-    if (!$scope.isActionHidden('lecturer-qa') || !$scope.isActionHidden('ai-qa')) {
-      $scope.rightBarTabs.questions = {
-        slug: 'questions',
-        title: 'Fragen',
-        disabled: false,
-        icon: 'glyphicon-pencil',
-        contentURL: 'partials/navBarRight/navBarRightQuestions',
-      };
-    }
-
-    // tab for code review
-    if (!$scope.isActionHidden('codeReview')) {
-      // check if the user already has a submission of the project or if the user is not of type `user` (e.g. `owner`)
-      const project = ProjectFactory.getProject();
-      const tabEnabled = project.projectCompleted || project.userRole !== 'user';
-
-      $scope.rightBarTabs.codeReview = {
-        slug: 'codeReview',
-        title: 'Code-Review',
-        disabled: !tabEnabled,
-        icon: 'glyphicon-eye-open',
-        contentURL: 'partials/navBarRight/navBarRightCodeReview',
-      };
-    }
-
-    /**
-     * tab for the sample solution
-     * sample solution will always be displayed if submission and solution is present in project
-     * therefore no isActionHidden check
-     */
-    if (!$scope.isActionHidden('sampleSolution')) {
-      // check if the tab should be enabled or not
-      const hasSampleSolution = ProjectFactory.hasSampleSolution();
-      const project = ProjectFactory.getProject();
-      const tabEnabled = hasSampleSolution && (project.projectCompleted || project.userRole !== 'user');
-
-      $scope.rightBarTabs.sampleSolution = {
-        slug: 'sampleSolution',
-        title: 'Lösung',
-        disabled: !tabEnabled,
-        icon: 'glyphicon-screenshot',
-        contentURL: 'partials/navBarRight/navBarRightSampleSolution',
-      };
-    }
-
-    // todo define other tabs
-
-    /**
-     * returns true when at least one tab is active
-     * @returns {boolean}
-     */
-    $scope.isNavBarRightActive = function () {
-      return !angular.equals({}, $scope.rightBarTabs);
-    };
-
-    /**
-     * check is tab is active
-     * @returns {boolean}
-     */
-    $scope.isTabActive = function (slug) {
-      return $scope.activeTab === slug;
-    };
-
-    /**
-     * Change content of tab splitter
-     * @param slug
-     */
-    $scope.rightBarTabClick = async function (slug) {
-      // indicate an event if the user has reached the request limit
-      // this is the case when the user has reached the request limit in a different tab
-      // and clicks on a new tab
-      if ($scope.isRequestLimitReached()) {
-        $rootScope.$broadcast(IdeMsgService.msgRequestLimitReached().msg);
-      }
-
-      if (slug === 'explanation') {
-        $rootScope.$broadcast(IdeMsgService.msgExpTabClicked().msg);
-      }
-
-      if ($scope.activeTab !== slug) {
-        $scope.splitter.expand('#ideRighterPartOfMiddlePart');
-        $scope.activeTab = slug;
-        TabService.setSlug(slug);
-
-        // remove 'tabClicked' class from all tabs on right side
-        for (var i = 0; i < tablinks.length; i++) {
-          tablinks[i].classList.remove('tabClicked');
-        }
-
-        // add 'tabClicked' class to the clicked tab on the right side
-        document.getElementById('rightBarTabs' + $scope.activeTab).classList.add('tabClicked');
-      } else {
-        $scope.splitter.collapse('#ideRighterPartOfMiddlePart');
-        $scope.activeTab = '';
-        TabService.setSlug('');
-        // remove 'tabClicked' class from all tabs on right side
-        for (var i = 0; i < tablinks.length; i++) {
-          tablinks[i].classList.remove('tabClicked');
-        }
-      }
-    };
-
-    /**
-     * toggle variable scope kPane
-     */
-    $rootScope.toggleVarScope = function () {
-      $scope.isCollapsed = !$scope.isCollapsed;
-      if ($scope.isCollapsed) {
-        $scope.innerSplitter.collapse('#ideVarScopePartOfMiddlePart');
-        $scope.ideTabsStyle = { 'margin-left': '47px' };
-      } else {
-        $scope.innerSplitter.expand('#ideVarScopePartOfMiddlePart');
-        $scope.ideTabsStyle = { 'margin-left': 'calc(10% + 47px)' };
-      }
-    };
-
-    /**
-     * function which gets called when user clicks on a new tab above the editor to remove the markers and close the window
-     */
-    $scope.$on(IdeMsgService.msgJavaTabOpened().msg, function () {
-      // reset storedMarkersBackup every time a new tab is clicked to not set the previous markers from the other tab in the new tab..
-      CodingAssistantCodeMatchSrv.storedMarkersBackup = [];
-      CodingAssistantCodeMatchSrv.toggleMarkers($scope.ace.editor, true, true);
-      if (!$scope.isCollapsed) {
-        $scope.innerSplitter.collapse('#ideVarScopePartOfMiddlePart');
-        $scope.ideTabsStyle = { 'margin-left': '47px' };
-        $scope.isCollapsed = true;
-      }
-    });
-
-    /**
-     * function which gets called when user clicks on the reset code btn to remove the markers and close the window
-     */
-    $scope.$on(IdeMsgService.msgResetCode().msg, function () {
-      CodingAssistantCodeMatchSrv.storedMarkersBackup = [];
-      CodingAssistantCodeMatchSrv.toggleMarkers($scope.ace.editor, true, true);
-      if (!$scope.isCollapsed) {
-        $scope.innerSplitter.collapse('#ideVarScopePartOfMiddlePart');
-        $scope.ideTabsStyle = { 'margin-left': '47px' };
-        $scope.isCollapsed = true;
-      }
-    });
-
-    /**
-     * function which gets called when user clicks on a new file in the tree-view to remove the markers and close the window - can be enabled in a case where the tree-view should be visible for the students
-     */
-    // $scope.$on('fileOpened', function () {
-    //     CodingAssistantCodeMatchSrv.toggleMarkers($scope.ace.editor, true, true);
-    //     if (!$scope.isCollapsed) {
-    //         $scope.innerSplitter.collapse('#ideVarScopePartOfMiddlePart');
-    //         $scope.ideTabsStyle = { 'margin-left': '47px' };
-    //         $scope.isCollapsed = true;
-    //     }
-    // });
-
-    /**
-     * function which gets called when code in ace editor changed to close the variable scope div
-     */
-    $scope.$on(IdeMsgService.msgCodeChanged().msg, function () {
-      if (!$scope.isCollapsed) {
-        $scope.innerSplitter.collapse('#ideVarScopePartOfMiddlePart');
-        $scope.ideTabsStyle = { 'margin-left': '47px' };
-        $scope.isCollapsed = true;
-        CodingAssistantCodeMatchSrv.toggleMarkers($scope.ace.editor, true);
-      }
-    });
-
-    /**
-     * This broadcast can be used to disable tabs from within a tab specific controller
-     */
-    $scope.$on(IdeMsgService.msgNavBarRightDisableTab().msg, function (event, data) {
-      $scope.rightBarTabs[data.slug].disabled = true;
-    });
-
-    /**
-     * This broadcast can be used to enable tabs from within a tab specific controller
-     */
-    $scope.$on(IdeMsgService.msgNavBarRightEnableTab().msg, function (event, data) {
-      $scope.rightBarTabs[data.slug].disabled = false;
-    });
-
-    /**
-     * Listens to open tab actions and opens the tab in the right bar.
-     */
-    $scope.$on(IdeMsgService.msgNavBarRightOpenTab().msg, function (event, data) {
-      $scope.splitter.expand('#ideRighterPartOfMiddlePart');
-      $scope.activeTab = data.tab;
-      for (var i = 0; i < tablinks.length; i++) {
-        tablinks[i].classList.remove('tabClicked');
-      }
-      document.getElementById('rightBarTabs' + $scope.activeTab).classList.add('tabClicked');
-    });
-  },
-]);
-
-// this service is used to make the clicked tab available in the ideNavBarRightCodingAssistantCtrl
-app.service('TabService', function () {
-  var slug;
-  var service = this;
-
-  service.getSlug = function () {
-    return slug;
-  };
-
-  service.setSlug = function (newSlug) {
-    slug = newSlug;
-  };
-});
-
-app.controller('IdeFooterStatusBarCtrl', [
-  '$scope',
-  '$routeParams',
-  'UserSrv',
-  'ProjectFactory',
-  function ($scope, $routeParams, UserSrv, ProjectFactory) {
-    /* Returns the username of the current user or '#anonymous' if user is not logged in */
-    $scope.getUsername = function () {
-      var _msg = 'User: ';
-
-      if (UserSrv.isAuthenticated()) {
-        _msg += UserSrv.getUsername();
-      } else {
-        _msg +=
-          '#anonymous (<a href="' +
-          $scope.signinSettings.signinPathWithRedirect() +
-          '">sign in</a> to save your progress)';
-      }
-
-      return _msg;
-    };
-
-    /* Returns a string that details the current user's role */
-    $scope.getCourse = function () {
-      return ProjectFactory.getProject().courseData ? ProjectFactory.getProject().courseData.coursename : '';
-    };
-
-    /* Returns a string that details the current user's role */
-    $scope.getRole = function () {
-      if ($scope.currentRoleIsLtiUser()) {
-        return 'LTI User';
-      } else if ($scope.currentRoleIsOwner()) {
-        return 'Project owner';
-      } else if ($scope.currentRoleIsUser()) {
-        return 'Project user';
-      } else if ($scope.currentRoleIsSubmission()) {
-        var _submissionRole = 'Inspection of a submission';
-
-        // we check we now the name of the user we're inspecting; if yes, we use the name as part of the role description
-        if (ProjectFactory.getProject().userBeingInspected) {
-          _submissionRole = 'Inspecting submission from user "' + ProjectFactory.getProject().userBeingInspected + '"';
-        }
-        return _submissionRole;
-      } else if ($scope.currentRoleIsUserProject()) {
-        var _userProjectRole = "Inspection of a user's project";
-
-        // we check if the url has parameter ?username=xxx; if yes, we use the name as part of the role description
-        if (ProjectFactory.getProject().userBeingInspected) {
-          _userProjectRole =
-            'Inspecting user-project from user "' + ProjectFactory.getProject().userBeingInspected + '"';
-        }
-        return _userProjectRole;
-      } else if ($scope.currentRoleIsHelp()) {
-        var _helpRequestRole = 'Inspection of a helprequest';
-
-        // we check we now the name of the user we're inspecting; if yes, we use the name as part of the role description
-        if (ProjectFactory.getProject().userBeingInspected) {
-          _helpRequestRole =
-            'Inspecting helpRequest from user "' + ProjectFactory.getProject().userBeingInspected + '"';
-        }
-        return _helpRequestRole;
-      }
-    };
-
-    /** Returns 'true' is the project is using Lti for the submission */
-    $scope.isUsingLti = function () {
-      return ProjectFactory.getProject().hasLtiData;
-    };
-
-    $scope.hasCourse = function () {
-      let courseData = ProjectFactory.getProject().courseData;
-      if (Object.keys(courseData).length === 0 && !$scope.currentRoleIsOwner()) {
-        $scope.disabledActions.submit = true;
-      }
-      return typeof courseData !== 'undefined';
-    };
   },
 ]);
