+++ app/views/partials/navBarRight/navBarRightInfo.html
@@ -1,19 +1,18 @@
 <div class="tab-body ide-tab-help ide-tab-chat" ng-controller="ideNavBarRightInfoCtrl" id="chat-wrapper">
-    <h1 class="modal-title">Erklärung der Helper-Systeme</h1>
-    <p>In diesem Tab werden dir die einzelnen <i>Helpersysteme</i> (rechte Navigationsleiste) sowie die <i>Buttons</i> (obere Navigationsleiste) für den Aufruf von weiteren Funktionalitäten erläutert.</p>
-    <hr class="chat-horizontal-line">
-    <div class="conversation">
-        <h2 class="info-heading">Helper-Systeme</h2>
-        <chat chat-lines="chatLines"></chat>
-    </div>
-    <hr class="chat-horizontal-line">
-    <h2 class="info-heading" ng-if="!isActionHidden('beautify') || !isActionHidden('varScope')">Buttons</h2>
-    <div class="button-row" ng-if="!isActionHidden('beautify')">
-        <button id="codeBeautifyButton" class="btn btn-info btn-sm"><span class="glyphicon glyphicon-flash"></span></button>
-        <span class="button-description">Nutze diesen Button, wenn du deinen Code formatieren willst</span>
-    </div>
-    <div class="button-row" ng-if="!isActionHidden('varScope')">
-        <button id="varScopeButton" class="btn btn-info btn-sm"><span class="glyphicon glyphicon-resize-vertical"></span></button>
-        <span class="button-description">Nutze diesen Button, wenn du den Gültigkeitsbereich von Variablen in deinem Code anzeigen willst</span>
-    </div>
+  <h1 class="modal-title">Erklärung der Helper-Systeme</h1>
+  <p>
+    In diesem Tab werden dir die einzelnen <i>Helpersysteme</i> (rechte Navigationsleiste) sowie die
+    <i>Buttons</i> (obere Navigationsleiste) für den Aufruf von weiteren Funktionalitäten erläutert.
+  </p>
+  <hr class="chat-horizontal-line" />
+  <div class="conversation">
+    <h2 class="info-heading">Helper-Systeme</h2>
+    <chat chat-lines="chatLines"></chat>
+  </div>
+  <hr class="chat-horizontal-line" />
+  <h2 class="info-heading" ng-if="!isActionHidden('beautify')">Weitere Tools</h2>
+  <div class="button-row" ng-if="!isActionHidden('beautify')">
+    <button id="codeBeautifyButton" class="btn btn-info btn-sm"><span class="glyphicon glyphicon-flash"></span></button>
+    <span class="button-description">Nutze diesen Button, wenn du deinen Code formatieren willst</span>
+  </div>
 </div>
